+++ x64barebones/Userland/SampleCodeModule/comandos/test_pipe.c
@@ -2,7 +2,7 @@
 #include "../include/syscalls.h"
 #include "../include/stdio.h"
 #include "../include/stdlib.h"
-
+#include "../include/test_util.h"
 
 static char * string = "This is a super long string and is supposed to test pipes. Write will write parts of the string to the pipe and read will read and check that its correct. All the while, we will be using busy waiting to make this process longer...";
 #define STRING_LENGTH 230
@@ -45,8 +45,8 @@ void test_pipe(){
 		return;
 	}
 
-	sys_register_child_process(&read_pipe, STDIN, NORMAL_SCREEN, NULL);
-	sys_register_child_process(&write_pipe, STDIN, NORMAL_SCREEN, NULL);
+	sys_register_child_process((uint64_t) &read_pipe, STDIN, NORMAL_SCREEN, NULL);
+	sys_register_child_process((uint64_t)&write_pipe, STDIN, NORMAL_SCREEN, NULL);
 
 	sys_wait_for_children();
